# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "School Management System API",
    "version": "1.0.0",
    "description": "Complete API documentation"
  },
  "paths": {
...
    "/api/v1/files/upload/single": {
      "post": {
        "summary": "Upload single file",
        "tags": [
          "Files"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "tenantHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "category": {
                    "type": "string",
                    "default": "test"
                  },
                  "isPublic": {
                    "type": "boolean",
                    "default": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/files/{publicId}": {
      "delete": {
        "summary": "Delete file",
        "tags": [
          "Files"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "publicId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File deleted successfully"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "FileUploadResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "File uploaded successfully"
          },
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "filename": {
                "type": "string"
              },
              "url": {
                "type": "string",
                "format": "uri"
              },
              "publicId": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "tenantHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Tenant-ID"
      }
    }
  },
...
}